requirements.txt
python-telegram-bot>=20.4,<21.0
numpy==1.23.5
gspread>=5.0.0
oauth2client>=4.1.3
python-dotenv>=0.19.0
pandas>=1.3.0